refactor: externalyze cubit state
